refactor: issue-#14,#15
